requirements.txt
requests
python-dotenv
prefect
tabulate==0.9.0
bs4
lxml
SQLAlchemy
psycopg2-binary